refactor: config and cli
