+++ src/config.rs
@@ -4,6 +4,7 @@ use crate::genesis::{GenesisRegistry, genesis};
 use clap::ValueEnum;
 use pallas_network::facades::NodeClient;
 use serde::{Deserialize, Serialize};
+use std::fmt::{self, Formatter};
 use tokio::runtime::Handle;
 use tracing::Level;
 
@@ -34,6 +35,16 @@ pub enum Mode {
     Full,
 }
 
+impl std::fmt::Display for Mode {
+    fn fmt(&self, f: &mut Formatter) -> fmt::Result {
+        match self {
+            Mode::Compact => write!(f, "compact"),
+            Mode::Light => write!(f, "light"),
+            Mode::Full => write!(f, "full"),
+        }
+    }
+}
+
 #[derive(Debug, Clone, ValueEnum, Serialize, Deserialize, PartialEq, Eq)]
 #[serde(rename_all = "lowercase")]
 pub enum Network {
@@ -50,6 +61,19 @@ pub enum LogLevel {
     Warn,
     Error,
     Trace,
+}
+
+// Implement conversion from LogLevel enum to tracing::Level
+impl From<LogLevel> for Level {
+    fn from(log_level: LogLevel) -> Self {
+        match log_level {
+            LogLevel::Debug => Level::DEBUG,
+            LogLevel::Info => Level::INFO,
+            LogLevel::Warn => Level::WARN,
+            LogLevel::Error => Level::ERROR,
+            LogLevel::Trace => Level::TRACE,
+        }
+    }
 }
 
 impl Config {
